+++ src/main/java/org/zalando/nakadi/controller/EventPublishingController.java
@@ -19,6 +19,7 @@ import org.zalando.nakadi.domain.EventPublishingStatus;
 import org.zalando.nakadi.exceptions.NakadiException;
 import org.zalando.nakadi.exceptions.NoSuchEventTypeException;
 import org.zalando.nakadi.exceptions.runtime.AccessDeniedException;
+import org.zalando.nakadi.exceptions.runtime.EventTypeTimeoutException;
 import org.zalando.nakadi.exceptions.runtime.ServiceTemporarilyUnavailableException;
 import org.zalando.nakadi.metrics.EventTypeMetricRegistry;
 import org.zalando.nakadi.metrics.EventTypeMetrics;
@@ -110,6 +111,9 @@ public class EventPublishingController {
         } catch (final NoSuchEventTypeException e) {
             LOG.debug("Event type not found.", e.getMessage());
             return create(e.asProblem(), nativeWebRequest);
+        } catch (final EventTypeTimeoutException e) {
+            LOG.debug("Failed to publish batch", e);
+            return create(Problem.valueOf(Response.Status.SERVICE_UNAVAILABLE, e.getMessage()), nativeWebRequest);
         } catch (final NakadiException e) {
             LOG.debug("Failed to publish batch", e);
             return create(e.asProblem(), nativeWebRequest);
