+++ src/main/java/org/zalando/nakadi/exceptions/runtime/EventTypeTimeoutException.java
@@ -1,0 +1,8 @@
+package org.zalando.nakadi.exceptions.runtime;
+
+public class EventTypeTimeoutException extends MyNakadiRuntimeException1 {
+
+    public EventTypeTimeoutException(final String message) {
+        super(message);
+    }
+}
